+++ src/index.ts
@@ -1,3 +1,8 @@
 export { Calendar } from "./components/calendar/calendar";
-export type { CalendarProps, StartOfWeek } from "./types/calendar";
+export type {
+  CalendarProps,
+  StartOfWeek,
+  CalendarMode,
+  DisabledRule,
+} from "./types/calendar";
 export type { CalendarTheme } from "./types/themes";